Refactor util file into folder
