+++ src/Components/AppPieces/UnauthorizedUser.js
@@ -4,7 +4,7 @@ import Typography from "@mui/material/Typography";
 import Stack from "@mui/material/Stack";
 import Button from "@mui/material/Button";
 
-import { getEmailLink } from "../../utils";
+import { getEmailLink } from "../../utils/values";
 import { emailConfig } from "../../constants";
 
 const getSendAuthorizationRequest = (user, addAlert) => () => {
